+++ src/core/context.ts
@@ -226,6 +226,7 @@ import {
     Line,
     map,
     mapc,
+    Mat2,
     Mat23,
     normalizedCurve,
     Point,
@@ -259,6 +260,7 @@ import {
     createStarPolygon,
 } from "../math/polygongeneration";
 import { insertionSort } from "../math/sort";
+import { makeQuadtree, Quadtree } from "../math/spatial/quadtree";
 import { Vec2 } from "../math/Vec2";
 import { BlendMode, type KAPLAYPlugin } from "../types";
 import {
@@ -333,6 +335,7 @@ export const createContext = (
         dt: app.dt,
         fixedDt: app.fixedDt,
         restDt: app.restDt,
+        setFixedSpeed: app.setFixedSpeed,
         time: app.time,
         screenshot: app.screenshot,
         screenshotToBlob: app.screenshotToBlob,
@@ -534,9 +537,12 @@ export const createContext = (
         Collision,
         Vec2,
         Color,
+        Mat2,
         Mat4,
         Mat23,
         Quad,
+        Quadtree,
+        makeQuadtree,
         RNG,
         Rule,
         RuleSystem,