+++ CHANGELOG.md
@@ -20,6 +20,9 @@ best friend, lajbel, can put the correct version name here
 
 ### Added
 
+- Added the `maxTimeStep` and `fixedUpdateMode` options, as well as
+  `setFixedSpeed()` for more granular control over fixed update and timing -
+  @dragoncoder047
 - Added parameterized formatting tags like `"[color=red]Red text![/color]"` in
   `CharTransformFunc` for more powerful text formatting options -
   @dragoncoder047
@@ -28,6 +31,10 @@ best friend, lajbel, can put the correct version name here
 - Added `createCogPolygon()` to create 2D regular cogs - @mflerackers
 - Added `getSpriteOutline()` that takes a sprite asset and returns a polygon
   showing the outline - @milosilo-dev
+- Added Quadtree for collision detection (only for fixed size screen for now,
+  needs expansion) - @mflerackers
+- Added vertical sweep and prune - @mflerackers
+- Added configuration to choose broad phase algorithm - @mflerackers
 
 ### Fixed